requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pytest==8.4.1
streamlit==1.49.0
torch==2.8.0
joblib==1.4.2
scikit-learn==1.6.1
pandas==2.2.3
altair==5.5.0
google-genai==1.47.0
python-dotenv==1.1.1

